requirements.txt
numpy==1.15.1
scipy==1.1.0
matplotlib==2.2.3
tqdm==4.26.0
python-speech-features==0.6
sox==1.3.3
scikit-learn==0.19.2
scikits.talkbox==0.2.5
torch==0.4.1
torchvision==0.2.1
baidu-aip==2.2.13.0